refactored serializers
