+++ api/serializers/__init__.py
@@ -1,0 +1,8 @@
+from .user_serializer import UserSerializer
+from .medication_serializer import MedicationSerializer
+from .patient_serializer import PatientSerializer
+from .visit_serializer import VisitSerializer
+from .vitals_serializer import VitalsSerializer
+from .order_serializer import OrderSerializer
+from .consult_serializer import ConsultSerializer
+from .diagnosis_serializer import DiagnosisSerializer
