+++ api/serializers/user_serializer.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from api import models
+
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = models.User
+        fields = ["username"]
